requirements.txt
transformers>=4.47.0
torch>=2.5.1
torchvision>=0.20.0
huggingface-hub>=0.25.1
sentencepiece>=0.2.0
jinja2>=3.1.4
pydantic>=2.9.2
timm>=1.0.9
tiktoken>=0.8.0
numpy==1.26.4
accelerate>=1.1.1
sentence_transformers>=3.1.1
gradio>=5.9.0
openai>=1.58.0
einops>=0.8.0
pillow>=10.4.0
sse-starlette>=2.1.3
bitsandbytes>=0.43.2
spaces>=0.31.1

# peft>=0.14.0 # Using with finetune model
